long ids

+++ Source/Podio .NET/Models/Action.cs
@@ -8,7 +8,7 @@ namespace PodioAPI.Models
     public class Action
     {
         [JsonProperty("action_id")]
-        public int? ActionId { get; set; }
+        public long? ActionId { get; set; }
 
         [JsonProperty("type")]
         public string Type { get; set; }
