+++ Source/Podio .NET/Models/Flow.cs
@@ -13,7 +13,7 @@ namespace PodioAPI.Models
         public string Name { get; set; }
 
         [JsonProperty("flow_id")]
-        public int FlowId { get; set; }
+        public long FlowId { get; set; }
 
         [JsonProperty("execution_count")]
         public int ExecutionCount { get; set; }
