+++ Source/Podio .NET/Models/Reminder.cs
@@ -5,7 +5,7 @@ namespace PodioAPI.Models
     public class Reminder
     {
         [JsonProperty("reminder_id", NullValueHandling = NullValueHandling.Ignore)]
-        public int ReminderId { get; private set; }
+        public long ReminderId { get; private set; }
 
         /// <summary>
         ///     Minutes to remind before the due_date
