+++ Source/Podio .NET/Models/Space.cs
@@ -6,7 +6,7 @@ namespace PodioAPI.Models
     public class Space
     {
         [JsonProperty("space_id")]
-        public int SpaceId { get; set; }
+        public long SpaceId { get; set; }
 
         [JsonProperty("name")]
         public string Name { get; set; }
@@ -21,7 +21,7 @@ namespace PodioAPI.Models
         public string UrlLabel { get; set; }
 
         [JsonProperty("org_id")]
-        public int? OrgId { get; set; }
+        public long? OrgId { get; set; }
 
         [JsonProperty("contact_count")]
         public int? ContactCount { get; set; }
